platform/api/requirements.txt
# ========================================================
# Python Backend Dependencies (FastAPI Clerk + Workers)
# ========================================================

# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0

# Data Ingestion (Project 1)
pandas>=2.2.0
openpyxl>=3.1.0  # For XLSX support

# Database
supabase>=2.4.0
asyncpg>=0.29.0          # Direct Postgres TCP pool — required for RLS (SET LOCAL session vars)

# Messaging (RabbitMQ)
pika>=1.3.0
aio-pika>=9.3.0

# Validation (Python 3.13 compatible)
pydantic>=2.10.0
email-validator>=2.1.0
httpx>=0.27.0


# SMTP Worker (Phase 3)
aiosmtplib>=3.0.0
bleach>=6.1.0
jinja2>=3.1.2

# Auth Security (Phase 1.5)
slowapi>=0.1.9          # Rate limiting per endpoint
python-jose[cryptography]>=3.3.0  # JWT encode/decode
bcrypt>=4.0.0  # Password hashing
python-multipart>=0.0.9 # Form parsing (FastAPI forms)

# Campaign State Management (Phase 4/5)
redis>=5.0.0
boto3==1.34.69
celery>=5.4.0
